+++ packages/open-truss/src/components/OTAvailableWorkflowsFromEndpoint.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import {
-  withChildren,
+  withChildrenV1,
   BaseOpenTrussComponentV1PropsShape,
 } from '../configuration'
 import { CSLinkShape } from '../shims'
 import { type z } from 'zod'
 
-export const Props = withChildren(BaseOpenTrussComponentV1PropsShape).extend({
+export const Props = withChildrenV1(BaseOpenTrussComponentV1PropsShape).extend({
   link: CSLinkShape,
 })
