+++ packages/open-truss/src/configuration/apply.tsx
@@ -1,9 +1,11 @@
 import { type z } from 'zod'
 import { OT_COMPONENTS } from '../components'
 import { type YamlObject, type YamlType } from '../utils/yaml'
-import { engineV1 } from './engine-v1'
-import { type WorkflowV1 } from './workflow-config'
-import { type BaseOpenTrussComponentV1 } from './components'
+import {
+  engineV1,
+  type WorkflowV1,
+  type BaseOpenTrussComponentV1,
+} from './engine-v1'
 
 export interface WorkflowSpec {
   workflow: WorkflowV1 // | WorkflowV2
