+++ packages/open-truss/src/configuration/engine-v1/index.ts
@@ -1,0 +1,9 @@
+// Ensure engine exports are namespaced with version to avoid collisions
+export { engineV1 } from './engine-v1'
+export {
+  BaseOpenTrussComponentV1PropsShape,
+  type BaseOpenTrussComponentV1Props,
+  type BaseOpenTrussComponentV1,
+  withChildrenV1,
+} from './components'
+export { type WorkflowV1 } from './workflow-config'
