+++ packages/open-truss/src/configuration/index.ts
@@ -1,7 +1,2 @@
 export * from './apply'
 export * from './engine-v1'
-export {
-  BaseOpenTrussComponentV1PropsShape,
-  type BaseOpenTrussComponentV1Props,
-  withChildren,
-} from './components'
